Remove chunk data serialization workaround to use serde serialization instead
