+++ core/anvil/Cargo.toml
@@ -16,7 +16,7 @@ thiserror = "1.0"
 tokio = { version = "0.2", features = ["full"] }
 serde = "1.0.112" # >= 1.0.112 needed for a fix in #[serde(flatten)]
 uuid = "0.8"
-hematite-nbt = "0.4"
+hematite-nbt = { git = "https://github.com/PistonDevelopers/hematite_nbt", rev="41124ff" } # needed for NBT array serialization
 byteorder = "1.3"
 bitvec = "0.17"
 anyhow = "1.0"
