+++ core/anvil/src/lib.rs
@@ -7,3 +7,4 @@ pub mod entity;
 pub mod level;
 pub mod player;
 pub mod region;
+mod serialization_helper;
