+++ core/anvil/src/region.rs
@@ -1,6 +1,7 @@
 //! This module implements the loading and saving
 //! of Anvil region files.
 
+use super::serialization_helper::packed_u9;
 use crate::{block_entity::BlockEntityData, entity::EntityData};
 use bitvec::{bitvec, vec::BitVec};
 use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
@@ -18,8 +19,6 @@ use std::io::{Cursor, SeekFrom};
 use std::ops::Deref;
 use std::path::PathBuf;
 use std::{fs, io, iter};
-
-mod blob;
 
 /// The length and width of a region, in chunks.
 const REGION_SIZE: usize = 32;
@@ -48,201 +47,58 @@ pub struct ChunkLevel {
     x_pos: i32,
     #[serde(rename = "zPos")]
     z_pos: i32,
+    last_update: i64,
+    inhabited_time: i64,
     sections: Vec<LevelSection>,
+    #[serde(serialize_with = "nbt::i32_array")]
     biomes: Vec<i32>,
     entities: Vec<EntityData>,
     #[serde(rename = "TileEntities")]
     block_entities: Vec<BlockEntityData>,
     heightmaps: Heightmaps,
+    #[serde(rename = "ToBeTicked")]
+    awaiting_block_updates: Vec<Vec<i16>>,
+    #[serde(rename = "LiquidsToBeTicked")]
+    awaiting_liquid_updates: Vec<Vec<i16>>,
+    post_processing: Vec<Vec<i16>>,
+    #[serde(rename = "TileTicks")]
+    scheduled_block_updates: Vec<ScheduledBlockUpdate>,
+    #[serde(rename = "LiquidTicks")]
+    scheduled_liquid_updates: Vec<ScheduledBlockUpdate>,
+    #[serde(rename = "Status")]
+    worldgen_status: Cow<'static, str>,
 }
 
+/// Represents the heightmap data of a chunk.
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "SCREAMING_SNAKE_CASE")]
 pub struct Heightmaps {
     // sometimes a few of those are missing, but we regenerate them anyways
+    /// Deserialization: length of 0 means field was missing,
+    /// length is guaranteed to be a multiple of 64,
+    /// length should be 256 to represent valid data
     #[serde(with = "packed_u9", default)]
     light_blocking: Vec<u16>,
+    /// Deserialization: length of 0 means field was missing,
+    /// length is guaranteed to be a multiple of 64,
+    /// length should be 256 to represent valid data
     #[serde(with = "packed_u9", default)]
     motion_blocking: Vec<u16>,
+    /// Deserialization: length of 0 means field was missing,
+    /// length is guaranteed to be a multiple of 64,
+    /// length should be 256 to represent valid data
     #[serde(with = "packed_u9", default)]
     motion_blocking_no_leaves: Vec<u16>,
+    /// Deserialization: length of 0 means field was missing,
+    /// length is guaranteed to be a multiple of 64,
+    /// length should be 256 to represent valid data
     #[serde(with = "packed_u9", default)]
     ocean_floor: Vec<u16>,
+    /// Deserialization: length of 0 means field was missing,
+    /// length is guaranteed to be a multiple of 64,
+    /// length should be 256 to represent valid data
     #[serde(with = "packed_u9", default)]
     world_surface: Vec<u16>,
-}
-
-impl Heightmaps {
-    // TODO remove this and use hematite-nbt serde for serialization
-    pub(crate) fn pack_u9(u9_array: &[u16]) -> Vec<i64> {
-        if u9_array.len() % 64 != 0 {
-            panic!("array length must be a multiple of 64");
-        }
-
-        let mut packed = Vec::with_capacity(u9_array.len() / 64 * 9);
-        let mut iter = u9_array.iter();
-
-        let mut container = 0u64;
-        let mut shift = 0;
-        for _elem in 0..u9_array.len() {
-            // For every element (u9)
-            let element = *iter.next().unwrap() as u64;
-
-            if element > 0x1FF {
-                // Invalid heightmap value
-                panic!("Invalid heightmap value {}", element);
-            }
-
-            container |= element << shift;
-            shift += 9;
-
-            if shift >= 64 {
-                // Take next container
-                packed.push(container as i64);
-
-                container = if shift > 64 {
-                    // We have some bits left to store in the new container
-                    element >> -(shift - 64 - 9)
-                } else {
-                    0
-                };
-
-                shift -= 64;
-            }
-        }
-
-        debug_assert_eq!(iter.next(), None);
-        debug_assert_eq!(shift, 0);
-        debug_assert_eq!(container, 0);
-
-        packed
-    }
-}
-
-mod packed_u9 {
-    use serde::de::Error as DeError;
-    use serde::de::{SeqAccess, Visitor};
-    use serde::export::Formatter;
-    use serde::ser::Error as SerError;
-    use serde::ser::SerializeSeq;
-    use serde::{Deserializer, Serializer};
-    use std::marker::PhantomData;
-
-    pub fn deserialize<'de, D>(deserializer: D) -> Result<Vec<u16>, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        struct PackedVisitor(PhantomData<fn() -> u16>);
-
-        impl<'de> Visitor<'de> for PackedVisitor {
-            type Value = Vec<u16>;
-
-            fn expecting(&self, formatter: &mut Formatter) -> std::fmt::Result {
-                formatter.write_str("a sequence of type long with length 36")
-            }
-
-            fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
-            where
-                A: SeqAccess<'de>,
-            {
-                let len = seq.size_hint().unwrap(); // nbt always knows sequence size
-                if len % 9 != 0 {
-                    // Invalid sequence length
-                    return Err(A::Error::custom("sequence length must be a multiple of 9"));
-                }
-                let unpacked_len = len * 64 / 9;
-
-                let mut u9_array: Vec<u16> = Vec::with_capacity(unpacked_len);
-
-                let mut container: Option<u64> = seq.next_element()?.map(|x: i64| x as u64); // We checked the length
-                let mut shift = 0;
-                for _elem in 0..unpacked_len {
-                    // For every element (u9)
-
-                    // unwrapping here is safe, as this can only fail if there is an implementation error in this algorithm
-                    // or in the SeqAccess because we checked the sequence length
-                    let mut element: u16 = ((container.unwrap() >> shift) & 0x1FF) as u16;
-                    shift += 9;
-
-                    if shift >= 64 {
-                        // Take next container
-                        container = seq.next_element()?.map(|x: i64| x as u64);
-
-                        if shift > 64 {
-                            // We have some bits left to get from the next container
-
-                            // same here with the unwrapping
-                            element |= ((container.unwrap() << -(shift - 64 - 9)) & 0x1FF) as u16;
-                        }
-
-                        shift -= 64;
-                    }
-
-                    u9_array.push(element);
-                }
-
-                debug_assert_eq!(container, None);
-                debug_assert_eq!(shift, 0);
-
-                Ok(u9_array)
-            }
-        }
-
-        deserializer.deserialize_seq(PackedVisitor(PhantomData))
-    }
-
-    pub fn serialize<S>(u9_array: &[u16], serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        // TODO use NBT LongArray type once implemented in hematite_nbt ( https://github.com/PistonDevelopers/hematite_nbt/pull/51)
-
-        if u9_array.len() % 64 != 0 {
-            // Invalid array length
-            return Err(S::Error::custom("array length must be a multiple of 64"));
-        }
-
-        let mut seq = serializer.serialize_seq(Some(36))?;
-        let mut iter = u9_array.iter();
-
-        let mut container = 0u64;
-        let mut shift = 0;
-        for _elem in 0..u9_array.len() {
-            // For every element (u9)
-            let element = *iter.next().unwrap() as u64;
-
-            if element > 0x1FF {
-                // Invalid heightmap value
-                return Err(S::Error::custom(format!(
-                    "invalid heightmap value {}",
-                    element
-                )));
-            }
-
-            container |= element << shift;
-            shift += 9;
-
-            if shift >= 64 {
-                // Take next container
-                seq.serialize_element(&(container as i64))?;
-
-                container = if shift > 64 {
-                    // We have some bits left to store in the new container
-                    element >> -(shift - 64 - 9)
-                } else {
-                    0
-                };
-
-                shift -= 64;
-            }
-        }
-
-        debug_assert_eq!(iter.next(), None);
-        debug_assert_eq!(shift, 0);
-        debug_assert_eq!(container, 0);
-
-        seq.end()
-    }
 }
 
 /// Represents a chunk section in a region file.
@@ -250,10 +106,12 @@ mod packed_u9 {
 #[serde(rename_all = "PascalCase")]
 pub struct LevelSection {
     y: i8,
-    #[serde(rename = "BlockStates")]
+    #[serde(serialize_with = "nbt::i64_array", rename = "BlockStates")]
     states: Vec<i64>,
     palette: Vec<LevelPaletteEntry>,
+    #[serde(serialize_with = "nbt::i8_array")]
     block_light: Vec<i8>,
+    #[serde(serialize_with = "nbt::i8_array")]
     sky_light: Vec<i8>,
 }
 
@@ -267,12 +125,33 @@ pub struct LevelPaletteEntry {
     properties: Option<LevelProperties>,
 }
 
-/// Represents the proprties for a palette entry.
+/// Represents the properties for a palette entry.
 #[derive(Serialize, Deserialize, Debug)]
 pub struct LevelProperties {
     /// Map containing a list of property names to values.
     #[serde(flatten)]
     props: BTreeMap<Cow<'static, str>, Cow<'static, str>>,
+}
+
+/// Represents a block update scheduled for a specific time.
+#[derive(Serialize, Deserialize, Debug)]
+pub struct ScheduledBlockUpdate {
+    /// The identifier of the type of this block
+    #[serde(rename = "i")]
+    name: Cow<'static, str>,
+    /// When this update should be executed in ticks from current time, can be negative if overdue
+    #[serde(rename = "t")]
+    ticks_from_now: i32,
+    /// Lower priority is handled first when happening on the same tick
+    #[serde(rename = "t")]
+    priority: i32,
+    // TODO are these global or chunk coordinates?
+    /// X coordinate
+    pub x: i32,
+    /// Y coordinate
+    pub y: i32,
+    /// Z coordinate
+    pub z: i32,
 }
 
 /// A block of sectors in a region file.
@@ -417,14 +296,11 @@ impl RegionHandle {
         // Write chunk to `ChunkRoot` tag.
         let root = chunk_to_chunk_root(chunk, entities, block_entities);
 
-        let blob = blob::chunk_root_to_blob(root);
-
         // Write to intermediate buffer, because we need to know the length.
         let mut buf = Vec::with_capacity(4096);
         buf.write_u8(2).map_err(Error::Io)?; // Compression type: zlib
 
-        blob.to_zlib_writer(&mut buf)
-            .expect("Could not write chunk blob");
+        nbt::to_zlib_writer(&mut buf, &root, None).map_err(Error::Nbt)?;
 
         let total_len = buf.len() + 4; // 4 bytes for length header
 
@@ -546,6 +422,8 @@ fn chunk_to_chunk_root(
         level: ChunkLevel {
             x_pos: chunk.position().x,
             z_pos: chunk.position().z,
+            last_update: 0,    // TODO
+            inhabited_time: 0, // TODO
             block_entities: block_entities.into(),
             sections: chunk
                 .sections()
@@ -592,6 +470,12 @@ fn chunk_to_chunk_root(
                     .map(|h| h.world_surface())
                     .collect(),
             },
+            awaiting_block_updates: vec![vec![]; 16], // TODO
+            awaiting_liquid_updates: vec![vec![]; 16], // TODO
+            scheduled_block_updates: vec![],          // TODO
+            scheduled_liquid_updates: vec![],
+            post_processing: vec![vec![]; 16],
+            worldgen_status: "postprocessed".into(),
         },
         data_version: DATA_VERSION,
     }
@@ -1058,97 +942,5 @@ mod tests {
                 count: 2
             }
         );
-    }
-
-    #[derive(Serialize, Deserialize, Debug, Eq, PartialEq)]
-    struct TestPackedU9 {
-        #[serde(with = "packed_u9")]
-        list: Vec<u16>,
-    }
-
-    #[test]
-    fn test_packed_u9_pattern() {
-        let data_u64 = iter::repeat(0xAAAA_AAAA_AAAA_AAAAu64 as i64); // 64-bit 0b1010...
-        let data_u9 = [0b010101010u16, 0b101010101u16].iter().cloned().cycle(); // corresponding 9-bit pattern
-
-        let unpacked: Vec<u16> = data_u9.take(256).collect();
-        let packed: Vec<i64> = data_u64.take(36).collect();
-
-        // Test blob serializer
-        let converted = Heightmaps::pack_u9(&unpacked);
-        assert_eq!(
-            converted, packed,
-            "is:        {:016X?}\nshould be: {:016X?}",
-            converted, packed
-        );
-
-        // Test serde serialization
-        let mut tokenized_vec = packed.iter().map(|&x| Token::I64(x)).collect();
-
-        let mut tokenized_sequence = Vec::new();
-        tokenized_sequence.push(Token::Struct {
-            name: "TestPackedU9",
-            len: 1,
-        });
-        tokenized_sequence.push(Token::Str("list"));
-        tokenized_sequence.push(Token::Seq { len: Some(36) });
-        tokenized_sequence.append(&mut tokenized_vec);
-        tokenized_sequence.push(Token::SeqEnd);
-        tokenized_sequence.push(Token::StructEnd);
-
-        let test_object = TestPackedU9 { list: unpacked };
-
-        serde_test::assert_tokens(&test_object, tokenized_sequence.as_slice())
-    }
-
-    #[test]
-    #[allow(clippy::inconsistent_digit_grouping)] // Sorry clippy but grouping by 9 bits makes sense here
-    fn test_packed_u9_order() {
-        let data_u64 = [
-            // this repeats every 9 u64...
-            0b0_001000000_000100000_000010000_000001000_000000100_000000010_000000001u64,
-            0b00_000100000_000010000_000001000_000000100_000000010_000000001_01000000u64,
-            0b000_000010000_000001000_000000100_000000010_000000001_010000000_0010000u64,
-            0b0000_000001000_000000100_000000010_000000001_010000000_001000000_000100u64,
-            0b01000_000000100_000000010_000000001_010000000_001000000_000100000_00001u64,
-            0b000100_000000010_000000001_010000000_001000000_000100000_000010000_0000u64,
-            0b0000010_000000001_010000000_001000000_000100000_000010000_000001000_000u64,
-            0b00000001_010000000_001000000_000100000_000010000_000001000_000000100_00u64,
-            0b010000000_001000000_000100000_000010000_000001000_000000100_000000010_0u64,
-        ]
-        .iter()
-        .cloned()
-        .map(|x| x as i64)
-        .cycle();
-        let data_u9 = (0..8).map(|x| 1 << x).cycle(); // corresponding 9-bit pattern
-
-        let unpacked: Vec<u16> = data_u9.take(256).collect();
-        let packed: Vec<i64> = data_u64.take(36).collect();
-
-        // Test blob serializer
-        let converted = Heightmaps::pack_u9(&unpacked);
-        assert_eq!(
-            converted, packed,
-            "is:        {:016X?}\nshould be: {:016X?}",
-            converted, packed
-        );
-
-        // Test serde serialization
-        let mut tokenized_vec = packed.iter().map(|&x| Token::I64(x)).collect();
-
-        let mut tokenized_sequence = Vec::new();
-        tokenized_sequence.push(Token::Struct {
-            name: "TestPackedU9",
-            len: 1,
-        });
-        tokenized_sequence.push(Token::Str("list"));
-        tokenized_sequence.push(Token::Seq { len: Some(36) });
-        tokenized_sequence.append(&mut tokenized_vec);
-        tokenized_sequence.push(Token::SeqEnd);
-        tokenized_sequence.push(Token::StructEnd);
-
-        let test_object = TestPackedU9 { list: unpacked };
-
-        serde_test::assert_tokens(&test_object, tokenized_sequence.as_slice())
     }
 }
